Refactor to separate the check job from the worker process. Also move more settings into ENV variables.
The worker process has been separated into worker.py so that the actual job now sits on it its own in check.py.
A new check_once.py has also been created which can be run if you want to do a single run of the check.
The Procfile has been updated so that Heroku now runs worker.py as the worker process.