Add a ttl for jobs when putting them on the worker queue
This uses the refresh rate as a time-to-live for jobs when they are queued - this means that a job will only ever wait in the queue for as long as it takes for the next job to be enqueued.
This will prevent a backlog of the same job building up if they are not being cleared out of the queue by the worker process for some reason.